+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 
-## [0.17.0] - 2023-05-16
+## [0.17.0] and [0.17.1] - 2023-05-16
 
 ### Added